+++ tailpos_sync/background_jobs.py
@@ -16,10 +16,17 @@ def generate_si_from_receipts():
     """
     pos_profile = frappe.db.get_single_value('Tail Settings', 'pos_profile')
     submit_invoice = frappe.db.get_single_value('Tail Settings', 'submit_invoice')
+    use_device_profile = frappe.db.get_single_value('Tail Settings', 'use_device_profile')
     company = frappe.db.get_value('POS Profile', pos_profile, 'company')
+
     receipts = frappe.get_all('Receipts', filters={'generated': 0})
 
     for receipt in receipts:
+        if use_device_profile:
+            device = frappe.db.get_value('Receipts', receipt.name, 'deviceid')
+            pos_profile = get_device_pos_profile(device)
+            company = frappe.db.get_value('POS Profile', pos_profile, 'company')
+
         si = frappe.get_doc({
             'doctype': 'Sales Invoice',
             'is_pos': 1,
@@ -49,7 +56,10 @@ def insert_invoice(invoice, submit=False):
     invoice.set_missing_values()
     invoice.insert()
 
-    invoice.payments[0].amount = invoice.outstanding_amount
+    invoice.append('payments', {
+        'mode_of_payment': 'Cash',
+        'amount': invoice.outstanding_amount
+    })
     invoice.save()
 
     if submit and not check_items_zero_qty(invoice.items):
@@ -60,3 +70,7 @@ def check_items_zero_qty(items):
     for item in items:
         if item.actual_qty <= 0:
             return True
+
+
+def get_device_pos_profile(device):
+    return frappe.db.get_value('Device', device, 'pos_profile')