+++ CHANGELOG.md
@@ -76,3 +76,4 @@
 
 ### v1.4.4 [10 мая 2019]
 - refactor: sync methods
+- feat: Device ID in syncing items